Fixed #21: Updated tests to be able to run as a GitHub action
