+++ Cartfile.resolved
@@ -1,1 +1,1 @@
-github "wultra/powerauth-mobile-sdk" "e5791d6bab0c643b4d0b62ebb71273938a13b7f9"
+github "wultra/powerauth-mobile-sdk" "e24aa6aaf4b12a2009eb02f6eb866549ac66f48e"
