+++ Tests/Configs/Readme.md
@@ -1,0 +1,13 @@
+To run `WultraSSLPinningTests` scheme you have to create a `config.json` file in this folder first with all the necessary information.
+
+_Example config:_
+
+```
+{
+    "url"           : "https://my-us-deployment.my-company.com",
+    "appName"       : "my-test-app",
+    "urlToPin"      : "https://github.com",
+    "adminLogin"    : "loginToAdminConsole",
+    "adminPassword" : "passwordToAdminConsole"
+}
+```
