--- cart-update.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-carthage update --use-xcframeworks --platform ios
